requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29
requests

# data science
pandas

# frontend
streamlit
python-dotenv
